refactor(auth): unify import statements and type usage in Login and Register
